Working on fixing the deletion form

+++ Calender/calender_delete_event.php
@@ -6,10 +6,14 @@
     if(isset($_GET['deleteuID'])){ 
         $uID = $_GET['deleteuID'];
         $ID = $_GET['eventID'];
-        $sql = "DELETE FROM `event` WHERE `ID`=$ID";
-        $stmt = $conn->prepare("DELETE FROM `event` WHERE `ID`=?");
-        $stmt->bind_param("s", $ID);
+        $stmt = $conn->prepare("DELETE FROM `event` WHERE `ID`=? AND `uID`=?");
+        $stmt->bind_param("si", $ID,$uID);
 
         $stmt->execute();
+
+        // $stmt = $conn->prepare("DELETE FROM `event` WHERE `ID`=?");
+        // $stmt->bind_param("s", $ID);
+
+        // $stmt->execute();
     }
 ?>
